+++ src/main/java/com/powsybl/openloadflow/knitro/solver/KnitroSolverFactory.java
@@ -75,7 +75,7 @@ public class KnitroSolverFactory implements AcSolverFactory {
         KnitroSolverParameters.SolverType knitroSolverType = knitroSolverParameters.getSolverType();
         return switch (knitroSolverType) {
             case STANDARD -> new KnitroSolver(network, knitroSolverParameters, equationSystem, j, targetVector, equationVector, parameters.isDetailedReport());
-            case RESILIENT -> new ResilientKnitroSolver(network, knitroSolverParameters, equationSystem, j, targetVector, equationVector, parameters.isDetailedReport());
+            case RELAXED -> new RelaxedKnitroSolver(network, knitroSolverParameters, equationSystem, j, targetVector, equationVector, parameters.isDetailedReport());
         };
     }
