+++ src/main/java/com/powsybl/openloadflow/knitro/solver/KnitroSolverParameters.java
@@ -279,6 +279,6 @@ public class KnitroSolverParameters implements AcSolverParameters {
 
     public enum SolverType {
         STANDARD,
-        RESILIENT
+        RELAXED
     }
 }
