--- src/main/java/com/powsybl/openloadflow/knitro/solver/KnitroSolverUtils.java
@@ -1,59 +1,0 @@
-/**
- * Copyright (c) 2025, Artelys (http://www.artelys.com/)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- * SPDX-License-Identifier: MPL-2.0
- */
-package com.powsybl.openloadflow.knitro.solver;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.List;
-import java.util.stream.IntStream;
-
-/**
- * @author Jeanne Archambault {@literal <jeanne.archambault at artelys.com>}
- */
-public final class KnitroSolverUtils {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(KnitroSolverUtils.class);
-
-    private KnitroSolverUtils() {
-
-    }
-
-    /**
-     * Builds the row and column index lists corresponding to the dense Jacobian structure,
-     * assuming each non-linear constraint is derived with respect to every variable.
-     *
-     * @param numVars               Total number of variables.
-     * @param listNonLinearConsts   List of non-linear constraint indices.
-     * @param listNonZerosCtsDense  Output list to receive constraint indices for non-zero Jacobian entries.
-     * @param listNonZerosVarsDense Output list to receive variable indices for non-zero Jacobian entries.
-     */
-    public static void buildDenseJacobianMatrix(
-            int numVars,
-            List<Integer> listNonLinearConsts,
-            List<Integer> listNonZerosCtsDense,
-            List<Integer> listNonZerosVarsDense) {
-
-        // Each non-linear constraint will have a partial derivative with respect to every variable
-        for (Integer constraintId : listNonLinearConsts) {
-            for (int varIndex = 0; varIndex < numVars; varIndex++) {
-                listNonZerosCtsDense.add(constraintId);
-            }
-        }
-
-        // We repeat the full list of variables for each non-linear constraint
-        List<Integer> variableIndices = IntStream.range(0, numVars).boxed().toList();
-        for (int i = 0; i < listNonLinearConsts.size(); i++) {
-            listNonZerosVarsDense.addAll(variableIndices);
-        }
-    }
-
-    public static void logKnitroStatus(KnitroStatus status) {
-        LOGGER.info("Knitro Status: {}", status);
-    }
-}
