+++ src/main/java/com/powsybl/openloadflow/knitro/solver/NonLinearExternalSolverUtils.java
@@ -11,13 +11,18 @@ import com.powsybl.openloadflow.ac.equations.AcVariableType;
 import com.powsybl.openloadflow.ac.equations.AcEquationType;
 import com.powsybl.openloadflow.equations.EquationTerm;
 import com.powsybl.openloadflow.equations.VariableEquationTerm;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.*;
+import java.util.stream.IntStream;
 
 /**
  * @author Jeanne Archambault {@literal <jeanne.archambault at artelys.com>}
  */
 public final class NonLinearExternalSolverUtils {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(NonLinearExternalSolverUtils.class);
 
     // List of always linear constraints
     private static final List<AcEquationType> LINEAR_CONSTRAINTS_TYPES = new ArrayList<>(Arrays.asList(
@@ -98,5 +103,38 @@ public final class NonLinearExternalSolverUtils {
             listCoef.add(scalar);
         }
         return new VarAndCoefList(listVar, listCoef);
+    }
+
+    /**
+     * Builds the row and column index lists corresponding to the dense Jacobian structure,
+     * assuming each non-linear constraint is derived with respect to every variable.
+     *
+     * @param numVars               Total number of variables.
+     * @param listNonLinearConsts   List of non-linear constraint indices.
+     * @param listNonZerosCtsDense  Output list to receive constraint indices for non-zero Jacobian entries.
+     * @param listNonZerosVarsDense Output list to receive variable indices for non-zero Jacobian entries.
+     */
+    public static void buildDenseJacobianMatrix(
+            int numVars,
+            List<Integer> listNonLinearConsts,
+            List<Integer> listNonZerosCtsDense,
+            List<Integer> listNonZerosVarsDense) {
+
+        // Each non-linear constraint will have a partial derivative with respect to every variable
+        for (Integer constraintId : listNonLinearConsts) {
+            for (int varIndex = 0; varIndex < numVars; varIndex++) {
+                listNonZerosCtsDense.add(constraintId);
+            }
+        }
+
+        // We repeat the full list of variables for each non-linear constraint
+        List<Integer> variableIndices = IntStream.range(0, numVars).boxed().toList();
+        for (int i = 0; i < listNonLinearConsts.size(); i++) {
+            listNonZerosVarsDense.addAll(variableIndices);
+        }
+    }
+
+    public static void logKnitroStatus(KnitroStatus status) {
+        LOGGER.info("Knitro Status: {}", status);
     }
 }
